Renaming main2
to just main.cpp
